Add handling of ModuleNotFoundError
